requirements.txt

Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.8.0
django-crispy-forms==1.10.0
djangorestframework==3.12.4
Pillow==8.0.1
requests==2.25.1
requests-oauthlib==1.3.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
